🚧 error converges at 26

+++ AI/314A3/parsedata.h
@@ -9,7 +9,7 @@ using namespace std;
 class Parser {
 public:    
     Parser();
-    pair<vector<vector<double>>, vector<int>> process_data(const string &filename);
+    pair<vector<vector<double>>, vector<double>> process_data(const string &filename);
 private:
     vector<string> split(const string &s, char delimiter);
 };
